fix: DEV-4081: e2e tests for selected region after tool changes (#1192)
* e2e tests to ensure selected region is correct after moving between certain tools.

When using Brush then Eraser or Pan, existing region should stay selected and have new additions be added to it. Similar for the Magic Wand.

We also have to ensure that the unselectRegionOnToolChange property for the Brush tool is updated on `afterCreate` so that e2e tests can dynamically update a feature flag and have it be applied.

* Make sure common helper function is available.

* Small code style fixes.

* Ensure magicwand tag is feature flagged on imports, not dynamically in the Registry.

This is safer for side effects. We've also modified our e2e tests to work with this different feature flagging approach.

* Make changes based on code review feedback.

---------

Co-authored-by: bradneuberg <bradneuberg@planet.com>
Co-authored-by: hlomzik <hlomzik@gmail.com>